add : AquaticAnimal class

+++ 13. Exams practice/Java OOP Retake Exam - 18 April 2022/01. High Quality Structure + 02. Business Logic/ZooApp/src/main/java/zoo/entities/animals/BaseAnimal.java
@@ -3,10 +3,10 @@ package zoo.entities.animals;
 public abstract class BaseAnimal implements Animal {
   private String name;
   private String kind;
-  private int kg;
+  private double kg;
   private double price;
 
-  protected BaseAnimal(String name, String kind, int kg, double price) {
+  protected BaseAnimal(String name, String kind, double kg, double price) {
     this.name = name;
     this.kind = kind;
     this.kg = kg;
